backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
langgraph
python-dotenv
msal
requests
pydantic
